+++ src/pages/product-registration/delivery/delivery.style.js
@@ -4,6 +4,7 @@ export const Container = styled.div`
   display: flex;
   flex-direction: column;
   width: 875px;
+  margin-bottom: 50px;
   border: 1px solid #eee;
 `;
 
@@ -31,4 +32,29 @@ export const RowTitle = styled.div`
 export const RowContent = styled.div`
   flex: 5.41;
   padding: 25px 30px;
+`;
+
+export const DatePickerContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding-top: 20px;
+`;
+
+export const DateTitle = styled.div`
+  font-size: 14px;
+  letter-spacing: 2px;
+  margin-right: 15px;
+  /* margin: 15px; */
+`;
+
+export const DateAndTitleContainer = styled(DatePickerContainer)`
+  padding-top: 0;
+  margin-right: 50px;
+`;
+
+export const Separator = styled.p`
+  display: flex;
+  justify-content: center;
+  margin: 0 20px;
 `;
